refactor: lights page simplify
