--- src/pages/lights/lights-schedule.module.css
@@ -1,21 +1,0 @@
-.root {
-	--gauge-color: gold;
-
-	display: grid;
-	justify-items: center;
-
-	header {
-		width: 100%;
-	}
-
-	p {
-		text-align: center;
-	}
-}
-
-.state {
-	display: grid;
-	grid-template: 1fr / auto 1fr auto;
-	justify-items: center;
-	margin-inline: 0 !important;
-}
