--- src/pages/lights/lights-today.module.css
@@ -1,35 +1,0 @@
-.root {
-	display: grid;
-	grid-auto-rows: min-content;
-	gap: var(--space-2);
-	padding-block-end: var(--space-6);
-
-	p {
-		justify-self: center;
-		font-size: var(--font-size-3);
-	}
-
-	& > svg {
-		height: 4rem;
-		justify-self: center;
-		margin-block-start: var(--space-2);
-	}
-}
-
-.state {
-	font-variation-settings: "wght" 500;
-	color: var(--color-danger);
-
-	&:global(.active) {
-		color: var(--color-success);
-	}
-}
-
-.note {
-	font-size: var(--font-size-2);
-	color: var(--color-text-secondary);
-	font-weight: 300;
-	margin-bottom: var(--space-2);
-	text-align: center;
-	text-wrap: balance;
-}
