+++ src/styles/utils.css
@@ -22,6 +22,7 @@
 		display: flex;
 		justify-content: center;
 		align-items: center;
+		width: 100%;
 		border-radius: 10px 10px 0 0;
 		padding: var(--space-2) var(--space-5);
